libft

+++ ft_strlcat.c
@@ -10,12 +10,12 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <libft.h>
+#include "libft.h"
 
 size_t	ft_strlcat(char *dst, const char *src, size_t size)
 {
-	int			i;
-	int			j;
+	size_t		i;
+	size_t		j;
 	size_t		srclen;
 	size_t		dstlen;
